Reverted component update logic to uninstall first
Turns out that installing the new version first then uninstalling the old version, when the update is simply a revision, causes the registration file for the packages in that version to vanish due to the fact that the uninstallation removes the registration file of that package name.

It can also cause installed files to be simply replaced then erased, leaving a broken installation.

So we made Component.CanInstall() public in order to allow checking before uninstalling.