feat: add .path units support for dotnet-installer
The .path units will be responsible for catching notifications that a content snap has been updated and triggering a service that updates all the mount units for said content snap.

This notification takes form of a file created in each content snap's own $SNAP_COMMON that the .path units will watch for.