fix: SnapInfo parsing for revison with non-numeric symbols
Parsing SnapInfo from json strings that contain revisons with
non-numeric symbols crashes, bcause a letter can not be converted
to int.

This commit chnages the datatype of SnapInfo.Revision to string.

An example for this seems to be installing a snap with the `--dangerous`
flag, which prepends an `x` (note: I could not find official
documentation for this).