feat: improve .NET version parsing
Currently the .NET version of the install and remove command
expected an "x[.x]" format. Further the ".x" part was never
verified and could have been arbitrary text.

This commit changes the version parsing to look for the pattern
`\A(?'major'\d+)(?:\.(?'minor'\d+))?\z` and verify that minor is 0
if specified.

Further the special values `latest` and `lts` are now changed
in a single place instead of implemented multiple times and
matching for these values is now case insensitive.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>