refactor: replace capture value empty check with capture success check
I misunderstood how `GroupCollection.ContainsKey` works. It even can throw
an exception when you wouldn't expect. For example: 
`DotnetVersionPattern.Match("a").Groups.ContainsKey("minor")` throws a 
`NullReferenceException`. That it throws an exception looks to me like a
bug in the library.

Let's replace `parsedVersion.Groups.ContainsKey("minor")` with
`parsedVersion.Groups["minor"].Success`, which actually does what we want.
`GroupCollection.Item[string]` always returns a value, even if the capture
group does not exist.

Also, checking again if the format of the parsed value is correct kind of
defeats the purpose of using a regular expression.